Added new abstract class
